OpenGL Context Setup and Regular Frame Cadence
Racket creates the OpenGL context via a racket/gui canvas% object, and makes
several FFI calls to a custom backend to finish initialization and facilitate
rendering.

As racket does not appear to have a mechanism for repainting an OpenGL canvas
every frame, a zero-delay timer% is used to provide the heart beat.  The OpenGL
context is set up to use vsync (with an unconfigurable interval of 1), and the
backend's Render callback issues glFinish before returning.  This results in
the correct throttling to ensure a cadence matching vsync's refresh rate.