Changed WGL Context Handling, Slightly Fancier Demo
Now there is a shader!

More importantly, when initializing the backend on windows, a new WGL context
is created independently from Racket's.  This uses the same HDC as Racket's
context, but passes in the attributes needed to request GL 4.2 Core.  This is
needed to be able to use Renderdoc for debugging.  Linux support will require
a followup change to do something similar for GLX.

As a result the new context handling, the backend now manages buffer swapping.
In a future change, I'd like to move the render loop into its own thread to
allow the overall UI to be more responsive.