Draw bounding box instead of full screen pass.
The most important thing about this change is that the depth value produced
upon rasterization and the depth value emitted by the fragment shader are
perfectly compatible, and so a future version of this could in theory mix
and match raster and raymarched content.

The bounding box draw produces both fewer discards, and has a much closer
ray start.

Ray misses write a depth value of 0, which since we're using inverse Z, places
the fragment on the far plane, effectively resulting in a discard.  The
depth_less qualifier is used in hopes that an early z test could early
eliminate fragments when the bounds would be occluded, but this does not
result in any useful optimization on the Nvidia hardware (a 1060) that I've
tested it on.  I'm not totally sure if the pairing of depth_less and GL_GREATER
ever does, though.

Performance wise, this exact approach is a dead end, as the fragment shader is
too expensive.  I think this might be salvaged by ditching the "sphere tracing"
ray marching algorithm when possible.  For example, a ray that only hits a box
should probably just use a ray/box intersection instead, or rasterization.
Some regions would still need to be raymarched, such as the various smooth
csg operators, but may be done so with much tighter bounds.  For example, the
smooth union operator only needs to trace the intersection of the two operands,
each padded by the maximum value of H, and the operands are rendered
separately, most likely discarding the intersecting region.