WIP Simple "Cluster" Culling
When this is complete, the raymarching draw will be spit up into tiles, and
draw from a heap of packets which is populated by a compute shader.

Refer to August 29th, 2021 journal entry for implementation notes.

The following steps are now implemented:

 1. Add an indirect draw params buffer.
 2. Convert the raymarch draw to an indirect draw.
 3. Add a compute pass to set the indirect draw params.
 4. Add a tile heap and metadata buffer.
 5. Change the compute pass have the generated AABB function.
 X. Unplanned, but necessary changes:
    a. Split the compute pass into two passes: cull and tally.
    b. Added a tile coverage visualization mode.

The following is mostly implemented, but is broken:

 6. The Compute pass performs aabb/tile hit test and only writes to the heap
    when it pass.

The following is unimplemented:

 7. Update the raymarch draw VS to read a tile from heap, and use the tile
    params to generate the position insetad of the scene AABB function.

 8. Store the shape AABB on the tile packet when it passes culling, update
    raymarch VS to read it from the tile info and pass it to the FS.

Once this is complete, this will provide a reasonable base to start working on
the CSG graph permutation switching and visbuffer occlusion culling.