New CSG Syntax For List Processing
This implements a new symbol list based system for building CSG trees.  The
high level usage is more or less the same, but the intermediary should be
easier to work with than the old struct based system.  This will be useful
for the graph permutation analysis stuff.

See journal entry for August 31st, 2021.