Shader Permutation Selection
This change changes shader generation to calculate a collection of useful
subtrees from the main CSG tree.  The cluster culling pass now uses this
to either eliminate a tile or generate the subtree index for that tile.

A future change will update the raymarching stage to use a generated
counterpart to raymarch only the tile's subtree.