Rudimentary Transform Coalescing
For each generated CSG tree, flatten out all transforms up to the first
non-transform node.  Then combine the adjacent translations, and the adjacent
rotations, and regenerate the tree.

This shaves about 3 ms off the frame time for the tire demo full screen on a
1060.  There's still successive transforms in other parts of the trees, so I
think this isn't quite the right approach.

I think I'm going to rewrite this later to work like so:

Recurse up the CSG tree. When a transform nests another transform, the should
be combined.  If the two transforms are the same kind, then consolidate them
as a new transform of the same type.  If the two transforms are not the same
type, then combine them as matrices.