Somewhat Better Transform Folding
Adjacent homogenous transforms are now combined, whereas before a csg operator
would terminate the folding.  This brings the draw time down to about 12 ms and
change on a 1060 in a full screen windows, according to nsight.

This change also introduces functions so that translations and rotations can
also be folded together, but there is a problem in the math somewhere, so this
isn't actually used at the moment.

It is also possible to fold through the csg operators, but this doesn't result
in any worthwhile performance improvement without being able to fold down into
a single transform (and might not either way), so currently this code does not
do that.

Coalesce.rkt has some commented out sections that could be used to easily restore
matrix folding and fold-through operators at a later date, once the kinks are
worked out.

The functions invert-mat4, quat->mat4, and mat* definitely need to be checked
for correctness.  It is also plausible that multiply order is wrong.  It is
unlikely that the majorness is wrong (this assumes column-major).

It might be better to just remove the invert-mat4 function, and instead generate
the matrixes from inverse transforms and combine appropriately.  I could add
a quat-inverse function, but tbh (transpose-mat4 (quat->mat4 some-quat)) is
probably good enough and is one less thing to vet for correctness.