Supporting Changes for Early Transform Folding
With these changes, it is now possible to coalesce transforms before calling
"segments" or "tree-aabb".  For whatever reason, if you actually do this right
now, the generated code is a bit slower.  I think this results in looser bounds,
but I haven't checked that closely.

So for now, I'm performing transforms after graph fracturing.

I'm keeping this around, because I think it might be worthwhile to try to split
the tree more aggressively on operators with the assumption that the bounds of
their children can't be transformed further by rotations below the operator.