Basic Dear ImGui Menu Bar
This has toggle full screen and exit implemented, and place holder menus for
opening and reloading files.

Dear ImGui does not provide an open file dialog widget, so I'll have to find
something else to provide that.  I'll probably use this:
https://github.com/mlabbe/nativefiledialog

After that, the next thing to do is to rework the racket side of tangerine to
be self contained as a racket library, which would ideally just be embedded
in generated.hpp.  From there, I guess scripts for individual models would then
be dynamically (re)required or eval'd or such, and I guess the output glsl
would be extracted from the underlying racket list, etc.  Ideally this would
eventually be ran in its own thread, and compiled to spirv maybe, and joins
with the render thread once it is finished.