Transform Coalescence Before CSG Tree Decomposition
Transforms are folded before running "segments" aka tree decomposition.  This
results in all transforms being folded up into a singular mat4 per brush.  This
resolves the issue where rotating an operator would screw up the apparent AABBs
of the operands.  The reason why this works is the brush's AABB is only rotated
once, instead of possibly applying a rotation on the results of say, a diff.

To make early coalescence viable at all, this change also adds an inverse
matrix function as well.  Otherwise nonsense AABBs would be produced.

Stepping through the results in renderdoc, there does seem to be a few draws in
the wheel that are expensive and don't seem to contribute much.  I'm unsure if
these are significant, but I plan to revisit the behavior of tree-aabb, since it
should be possible to produce much tighter behavior for operations like diff
now.