Revised 'diff' Behavior for More Aggressive Splitting
Prior to this change, the following would result in a single AABB region:

    (diff (cube 4)
          (move-x 3.5 (cube 3.9))
          (move-x -3.5 (cube 3.9)))

With this change, the above now produces a cut for each rhs operand, and
a set of bounds for the remainder, which should be a lot faster to render.

This approach should eventually be applied to the intersection and blend
operators as well.  However, as implemented there is some room to improve.
Specifically, I'd like it to handle this case more gracefully:

    (diff (cube 4)
          (box 5 2 2)
          (box 2 5 2)
          (box 2 2 5))

In this case, the intersection between the three boxes should be a bound,
and the non-intersecting areas should be different bounds.  Currently these
just end up with the same subtree.  This might be fine anyway.