The Even More Aggressive Splitting for 'diff' Operations
Prior to this change, the following would result in a single AABB region:

    (diff (cube 4)
          (box 5 2 2)
          (box 2 5 2)
          (box 2 2 5))

Now it will produce 1 tightly bound cut for the overlapping region of all three
boxes, 2 bounds per box for the non-overlapping regions, and 20 bounds for the
portions of the cube which do not overlap with the boxes.

I think the basic idea here should also be applicable to the intersection
operator as well as the blend operators.  This should be useful for creating
repeated textures with cuts or unions.