Consolidate Compiled Shaders Into Low Level Model Object
This change reworks shader compiling and management to be consolidated in a
model object.  This will make it a lot easier to later implement the ability
to render multiple models at once.

LowLevelSubtree represents a single subtree within a model.  This struct wraps
the shader pair needed to render it, the generated sources needed to compile
it, and the number of AABBs that need to be drawn.  LowLevelSubtree provides
the boilerplate to compile the shader pair it wraps.

LowLevelModel builds a vector of LowLevelSubtree entries, and provides a simple
wrapper to compile the entire model.

Currently, there is a LowLevelModel pointer for the model to be rendered
(CurrentModel), and one for a pending model to replace it (PendingModel).

To implement hierarchical models and instancing from here, the following would
be needed:

 1. An API for racket models to call to instance other models, possibly without
    emitting a CSG tree.

 2. Rework loading machinery to have an explicit distinction between reloading
    vs redundant loading calls.

 3. Loaded LowLevelModels stored in a map keyed off of absolute path.

 4. A model instance object, and a list of live instances per model.

 5. The cluster culling CS will need to be dispatched with the
    instance count as the group z count, and probably need to write the
    instance ID into the tile heap so that the depth pass can use it.

 6. Each instance will need a matrix and its inverse to convert between local
    space and world space.  The natural thing to do here is have RenderFrame
    upload all of them into a single buffer, and then the cluster culling
    dispatches will need the appropriate offset.