Culling Setup Pass
This change moves most of the math out of cluster_cull.cs.glsl into the new
shader cluster_setup.cs.glsl, as part of my new "don't dispatch 55 million
threads when there's 100 instances" initiative.

This improves performance quite a bit.  I intend to follow this up with a change
to have the setup stage set the indirect dispatch arguments for the culling
stage.