Removed Placeholder Buffer Binding
Both indirect dispatch schemes I had in mind didn't work out to be faster than
just spawing a thread per bound per instance per screen tile, but moving most
of the ALU work to the setup stage was itself a significant optimization, so,
I think I'm going to just leave it as is for now.

I'm taking this to mean that the atomic add in the culling stage is where it's
bottlenecked.  If that is correct, then implementing occlusion culling should
also show a reduction in the overall time spent culling in addition to the
expected gains in the depth draw stage.