WIP Occlusion Culling
This implements the depth range history thing, and depth base tile elimination.
When enable, rotating the camera will cause tiles to sometimes momentarily
disappear as they fail culling.  This also changes the behavior of the clip
rect ranges producet by the setup stage, so that the z value is compatible with
the value stored in the depth buffer instead of the actual clip space value.

This is off right now as there are a few more things to do:

 1. The culling behavior is wrong at some angles when there are multiple
    instaces.  This doesn't seem to be due to the extra condition in the
    culling shader so much as the first test in ClipTest is correct now, and is
    revealing an issue I missed before.  Seeing as the problem only shows up
    with multiple instances, I think it might be something sketchy in the
    mapping between setup's writes and cull's reads.

 2. Set of culling and draw passes is needed, as is another gather after the
    first set of draws.  I think the culling setup results can be reused.  The
    second culling stage should eliminate everything that passed the first
    stage.

It may be that this culling strategy is too fine, and that I would get better
gains culling the AABBs against an HZB.  Once the first item above is
corrected, if attempting to cull 100 instances of objects is still super
expensive, then that may be a good sign that step 2 isn't worth implementing.

If I do switch to an AABB based strategy, then it may not be useful to continue
using a tiled renderer and may be better to switch back to a bounding polygon
variant.  Or stay with a tiled system, but do the tile elimination in a VS,
which would maybe eliminate the need for the atomic and the tile heap.