Tileless Rendering / New Instancing System
This restores some of the behavior from e5516159d22ce5e9c136ba56770f22af613e5f33,
where AABBs were just drawn as bounding boxes instead of the later tile based
system, and deletes said tile based system.

This is moving towards a setup where the brush transforms and parameters are
totaly decoupled from the compiled shader, but it is not quite there yet.  The
essential ground work is already in place though.  To complete this, the racket
frontend will need to be updated to extract parameters and transforms, revise
the generated GLSL, and prepare the data to be uploaded.

Beyond that, while Tangerine is mostly at feature parity to where it was
before, there is a problem where the instances aren't managed right, so things
get weird when you reopen a model.  Ideally the shader permutations will stay
open, but the associated instances should be reset and re-uploaded.