Reworked Diff Behavior
If you make a giant pile of bricks, such as in the bricks demo, and then
subtract a large shape from it, this will compile quickly and run fast now
instead of not doing those things.

Somewhere along the way it looks like the "cut a lot of things from something"
demos got a lot slower, but I think the regression was some time prior to this
change.