CSGST Material Annotations
These don't do anything yet, but will eventually be used for describing how
brushes should be rendered.  The material parameter to the "paint" and
"paint-over" functions will probably end up being a slot number or a semantic
name.  This will need to be paired with some additional information that would
likely be provided by the compile function.