Material Rendering Prototype
With this change, the "paint" nodes actually do something!  These are the
current available materials:

 0 - default
 1 - unlit black
 2 - unlit gray
 3 - unlit white
 4 - unlit red
 5 - unlit green
 6 - unlit blue

These are just test materials.  In the future these will not be hard coded.
The material passes render into a RGB8 buffer, using a specially constructed
depth buffer to function as a stencil w/ the GL_EQUAL depth test.  This is
intended to provide a higher range of values than is possible with a stencil
test.  Once the material passes have been drawn, the color buffer they write
to is resolved into the back buffer.  In the future, I'd like to change the
color buffer to be RGB32F in linear color, and have the final resolve perform
the appropriate tonemapping etc.