C++ Evaluable CSG Trees
This change implements a system for constructing SDF trees for the purpose of
evaluating them on the CPU.  This code is not currently in use by anything.

The intended future use is Racket uses the exposed C API to contstruct SDF
Trees from CSGST expressions after performing transform folding.  The tree root
would be given to the C++ application to use for tasks like mesh generation.
In the future, tree evaluation is also likely to be used as part of a voxel
based CSG tree decomposition algorithm.