SDF Evaluator Construction
This change implements the Racket machinery for translating CSGST expressions
to equivalent SDF evaluator node trees.  This also introduces the concept of
the active evaluator.  The active evaluator (global `TreeEvaluator`) is created
every time `compile` is called.  When a new one is set, the old one is
automatically deleted.  Additionally, the active evaluator is deleted when
loading a model before calling `emit-glsl`.

The active evaluator is not currently used for anything, but is intended to
be used for a future mesh export mode.