Export to Binary STL Mesh
This implements a working mesh exporter, though it still needs to be hooked up
to the UI.  The mesh is voxelized at an arbitrary resolution and then smoothed
along the distance field.  Both steps can take a long time for larger and/or
complex models, so this also parallelizes what it can.  This gets the job done,
but it runs long enough that I'll want to add a progress bar.