Eval Graph Node Bounds WIP
This adds boundary information to the SDF evaluator graph used by mesh export.
The intention is for it to be possible to accurately evaluate the distance
field without evaluating the entire graph, but it hits a bit short of the mark.
The "Clip" method is meant to accomplish this, but when used, the mesh export
is corrupted for more complex models, so its usage is disabled right now.

I'm considering a variant where instead the Clip method returns a new SDF tree
which only includes nodes reachable for a given volume, at which point Eval and
Gradient would be safely constrained, and I wouldn't have to worry about infs.