SDF Evaluator Trimming
This change implements a way to reduce SDF evaluator graphs to a locally
relevant subset.  With this change, SDFNode::Clip is called with a point and a
radius, and returns either a new graph evaluator, or nullptr.  The Clip method
evaluates the graph, and discards all nodes which cannot contribute a hit
within the clipping sphere.

This is useful in situations where the graph is expected to be called many
times within a small region of space.  This change uses it to greatly speed up
the refinement export step.

This might be later used to great effect in the mesh generation step with some
sort of octree like algorithm guiding successive graph clipping.