Simpler Model Boundary Estimation
This change adds a boundary estimation function that works by traversing the
csg tree up to the brushes, and then merging and transforming the bounding
boxes back to the root.  This differs from the current strategy which merges
the results of decomposition at the end.

This is intended for determining the volume limits needed for the upcoming eval
compiler.