Voxel Compiler
This change adds a new shape compiler called the "voxel-compiler".  The old
compiler is now called the "bounds-compiler".  This new compiler performs much
better with models like the gear model, but performs much worse with the brick
wall models.  Performance degrades rapidly with larger voxel sizes as the
compiled subtrees become larger than would be ideal.  There isn't a good way
to determine the ideal voxel size up front, so it is just .25 by default now.

This does fix the z-fighting problems of the bounds compiler and produces more
correct results, so I'm going with this as the default for now.  I'm going to
explore a binary searching variant where the voxel sizes don't need to be fixed.

There's also two issues with the evaluator tree system that affect the voxel
compiler.  First is the paint elements are discarded, and second is that clip
will include fully occluded volumes.  Both of these should be easy to fix.