Async GLSL Compiling
This change reworks the low level shader compiler to run asynchronously with
the renderer.  When the Racket frontend finishes processing the model, the
application becomes responsive again.  Every frame after until all jobs have
finished, some pending jobs are started, and corresponding drawables are
immediately created.  A drawable will be only drawn if it is valid and the
shader compiler job has finished.  The "OpenGL" model loading perf stat now
indicates the convergence time.

There are a few areas that need additional work:

 1. Starting a shader compiler is most likely IO bound.  This could be greatly
    improved by caching shader sources read from disk.  Better yet, the
    generated code could be compiled in a separate shader object and linked
    with a cached shader the static sources.

 2. There's no cancel mechanism, and it is plausible that loading new models
    can leak memory.

 3. It would be nice to have a fallback shader for rendering the voxels
    corresponding to in-progress shaders.

 4. In retrospect it was probably overly optimistic of me to express the
    loading stats in milliseconds, when the time scale is seconds.

As for performance, this is a significant improvement, as the amount of frame
time in which the renderer is blocked on the OpenGL shader compiler is
effectively zero now.  The convergence time is significantly lower than the
previous blocking time, and addressing point 1 above will improve it further.

The Racket frontend is now the primary load time bottleneck, as it blocks the
renderer.  If it were made to be asynchronous, the frontend is not currently
designed to emit backend jobs until after most of the processing is complete,
and if it could, merging drawables would be a problem.  I think these problems
can be overcome though without changing the high level design significantly.

It would be interesting if the compiler could be made to run continuously with
a caching / merging feature, where the steps of descent into the volume is
determined in part by expected screen space size and visibility, and to allow
for live changes to the model.  Drawables could also probably be handled
differently now since they're effectively sets of voxels now.