Moved Bounds Calculation Into C++ Backend
In the tiles.rkt test case, the "model bounds" time was about 35.56 ms prior to
this change, and is about 0.034 ms now.  The new bounds calculation is
essentially the same algorithm, just written in C++, and doesn't need to
dispatch CSGST types because it's just descending the SDF tree.

This change removes SetLimitsCallback, and it is replaced by SetTreeEvaluator
which now just recalculates the bounds again at the end (this should eventually
be consolidated further, probably).

This change merges EmitSubtree with EmitShader, since they are always called in
tandem.  This results in a small savings for more complex models, as FFI calls
are not free.

I'd like eventually for the compiler frontend to only run the input model to
produce the CSGST tree to construct the SDF evaluator directly, and perform
everything else in the C++ backend, including transform folding.