Reimplemented Voxel Compiler in C++
This removes most of "voxel-compiler" and everything after that in the Racket
frontend and replaces it with a faster C++ implementation.  This saved about
~280 ms in the "tiles.rkt" test case.

There's a bit of cleanup that needs to be done now, there are some obvious
areas for improvement, and I haven't exhaustively tested that everything still
works.

The main things I'd like to do soon are:

 1. Fix the debug names for subtrees / shader variants.

 2. Use instancing to draw parameter instances.

 3. Clean up the renderer to use consistent terminology / general refactoring.
    There's some redundant maps for deduplicating subtrees that could be
    consolidated.

 4. Prune out dead code.