Simplified Shader Compiling, Removed Async
glValidateProgramPipeline was producing a stall, which was probably a linking
stall.  Because GL_ARB_parallel_shader_compile mentions no equivalent for
COMPLETION_STATUS_ARB for program pipelines, and because I was not using
program pipelines in any meaningful way, I decided to simplify.

Unfortunately querying COMPLETION_STATUS_ARB to check a program object's
linking status *always* reports true, effectively rendering this extension
useless.

Because of the above, this commit rips out all usage of
GL_ARB_parallel_shader_compile, which makes convergence time considerably
worse.

I will be exploring a few other options:

First is to create another GL context in another thread for compiling shaders.
Assuming it is not possible to just transfer the compiled program object back
to the main context, this could be used either to warm up the driver's shader
cache before compiling again on the main context.  This should be viable for
atleast NVidia, but is unknown for other IHVs.

Second is to create another GL context in another thread like before, but save
out the program binary when finished and load it again on the main thread.
The lore recommends rendering with such shaders a few times before saving them
out to ensure background optimizations happen.  Doing such a thing would not
fit, so this might end up producing slower shaders.

Third option is switch to spirv and use glslang to compile.  This would allow
more control over optimization.  Hopefully the result is a fast translation
to a program binary, but idk what to realistically expect here in terms of
stalls.