Partially Working Async Shader Compiling
I went and reworked context creation to call wgl directly instead of trying to
do it with SDL.  Seems to mostly work.  The initial load times feel longer
somehow, and the convergence time stat needs to be reworked.  Sometimes the
models don't finish loading, and its possible to crash the driver still.

I was able to hit the driver crash trying to reload a model that seemed stuck,
which makes me think I probably deleted the program object while it was linking
or something.

Feels promising, but not ready to be merged.