Attempting to Improve Async Compile Perf
I think this is an improvement in as far as the renderer doesn't totally
deadlock now half way through opening the wheel model.  However, it is
only crawling along at 1 hz :/

I'm considering dropping the sharedness of the worker thread contexts in favor
of extracting the program binary and passing that back to the main thread.

Before I do that I should get debug messages working again(?) and see if there
are any warnings from the driver.

When it was deadlocking, the main thread would often get stuck in a random GL
get call in the dearimgui backend, while the worker threads would be waiting
on compile statuses.