Replaced Incomplete Material System With Simple Color Annotations
The paint function now accepts a color tripple, a HTML-style hex color code, or
a CSS color name in the form of a symbol or a string.

The "depth" pass will write this color into the material buffer instead of a
material integer ID.  The Outliner pass now reads this buffer to produce the
base color before applying the outline and simple shading.

This should make it much easier to produce color model exports eventually, and
is a much friendlier interface for defining basic coloration.

At some point in the future a more complicated system may be called for, but
seeing as there's no lighting model yet, there's no need for such a thing now.