Reduce Clipping False Positives
This adds an extra test to clip set nodes based on the local distance field
before checking the child nodes.  This eliminates some invisible false
positives that the previous behavior couldn't account for.

This doesn't catch everything however.  It may be due to inexact interior
bounds, but negative space cuts seem to also match voxels along the cut shape's
isosurface if the surface is close to the original surface.  This can be worked
around by ensuring the cut shape is sufficiently large, but I'm not ok with
having invisible perf footguns.

Removing the remainder of the false positives will take some thought, but if
nothing else, it looks like solving the infinitely thin surface problem will
probably also solve this one.