Remove PBuffer Creation Path For Shared Contexts, Disable Async Compiling
The PBuffer path never worked, which effectively disabled async shader
compiling anyway.

It technically works, but the driver is locking the contexts for some reason
and I don't think there is any way to determine why.

I'm planning on ripping this out and replacing it with SPIRV at a later date.
A secondary context might still be needed for a priming draw, but that can be
addressed once it is proven to be necessary.