Octree Coalescence
This change adds the ability to compare evaluator trees for equivalence.

This change also updates SDFOctree to apply this new functionality with a
coalescence feature.  When an octree's children are all termini or null, and
all share the same evaluator tree as the parent, then the parent is marked as a
new terminus, the bounding box is replaced with the combined AABB from the live
children, and the children are deleted.

For my WIP "cafe" model, this drops the total depth draw time from ~10ms down
to about ~5ms (RTX 2080 Ti).

This can be disabled via the ENABLE_OCTREE_COALESCENCE define for debugging,
see sdfs.h.