CSG Leaf Count Visualizer
This is to help visualize how effective the octree compiler is at reducing CSG
trees.  This is not a proxy for performance.  Trees containing many nodes can
often be cheap depending on a variety of factors, and the reverse is also true.

For clarity, this change also replaces "Complexity" with "LeafCount" in the sdf
evaluator sources.

I'd like to eventually follow this up with a mouseover tooltip that shows the
exact count, as well as the reduced CSG tree.