Increase Stack Reserve Size In Debug Builds to 10 MB
The color cube model was hitting a stack overflow in debug builds only while
calculating the model's full bounding box.  This was due to the Evaluator's
excessive tree depth.

This change is meant to compensate for the extra memory overhead of debug
builds, and is not a solution that should be applied to release builds.

In theory, any model that applies a set operator to enough operands will trip
this in any build, and the real solution to this problem will probably require
building some kind of stack machine instead of recursion for the main evaluator
methods.