Added Infinite Plane Operand
The infinite plane operand behaves like a brush shape, but it has an infinite
bounding area.  Because of the potential for NaNs and horrible infinite loops
that this introduces, Tangerine will now throw an error if you attempt to
build an evaluator with an infinite area.  This effectively means that planes
are only useable in the diff and inter operators.

The evaluator itself will work just fine with an infinite boundary, but shader
compiling and model export both require finite bounds.  Should it be useful,
the finite requirement might be loosened a bit in the future.

The infinite plane operand has a special case where when it is parallel to one
of the axes, the appropriate AABB min or max channel will be set to zero,
which could be used for constructing finite boxes.  However, you would be
better served by the box brush.