Tangerine Shader Interpreter!
This change implements an interpreter mode for Tangerine.  When the commandline
arg "--interpreted" is provided, all voxels will generate the same GLSL, which
opts into the new interpreter.  This is a hack to make an interpreter mode
available to accelerate offline rendering.  I plan to update this so all voxels
still generate the GLSL they normally would, and instead use the interpreter
automatically while shaders are compiling.

Right now an arbitrary stack size is provided with no means of changing it.
Surprisingly this works fine with all of my test models so far.

Also surprising is that the perf is not nearly as bad as you would expect from
an early implementation!