Speculative Interpreter Bug Fixes
This change fixes two hypothetical bugs in the interpreter shader.

The first bug is that malformed input could cause the stack pointer to run out
of bounds, most likely resulting in a TDR.  To prevent this, the interpreter
will return an error value if the stack pointer is out of bounds before
dispatching the next opcode.

The second bug is that a model mixing shapes with and without paint annotations
might accidentally pick up a stale paint annotation for an unannotated brush.
To prevent this, the color field is cleared every time the stack pointer is
pushed.