Rebalance Union and Intersection Ops to Improve Interpreter Perf
This change updates SetNode to swap the operands on construction when possible
to ensure the tree is left-leaning.  This is beneficial for the interpreter's
startup and steady state performance, as it can reduce the required stack size.

With this change, the wheel model dropped from a stack size of 6 to 4.