Stack Optimizing Operand Swaps
This changes operand swapping from occuring based on the leaf count to the
stack size.

This is effective at further reducing the required stack size needed for the
interpreter mode, and also helps consolidate shader permutations a little in
the compiled mode.

The gains to this are marginal, as it will only show up in models that would
otherwise run at a larger stack size.  Notably spiral.rkt is down to 3 now,
whereas it had a stack size of 4 with the leaf count based swaps, and a stack
size of 20 before that (these figures are maximums).  However this means
rounding stack sizes is unlikely to be useful.

Further optimizations to the interpreter are unlikely to be meaningful until
I implement occlusion culling, as octree coalescence makes profiling this sort
of change unreliable.