Remove Potential LGPL Dependency
As of Racket CS 8.5, the readme in Racket's toplevel directory states that all
packages included with Racket have their own licenses, and that one should
check the individual packages in Racket's "share/pkgs" folder.  Unfortunately
at this time of writting, the "math-lib" package is still licensed under the
LPGL: https://github.com/racket/math/blob/47b6122fe7979af80017ef3ba2b7cefc7514c1a3/math-lib/LICENSE.txt

Several people involved in the project believe this is a mistake, and hopefully
they are correct and this can be rectified in a few days.

However, "maybe" is not legally binding, and it is unknown how long this will
take to correct.

To be 100% sure, this change removes all direct dependencies to math-lib.

Tangerine only trivially used Racket's math libraries for:

 1. Converting from degrees to radians.

 2. Converting numeric types to single precision floating point numbers.

With this change, Tangerine no longer has any direct dependencies to Racket's
math libraries.

Additional work will be needed to audit the dependency hierachry to determine
Racket's suitability for further use in this project.