Revised Paint Application Semantics
This commit changes the behavior of the intersection operator, so that when one
operand has a material and the other does not, the present material is applied
to both operands.

In the interpreter, a null material is indicated via a negative color.  In the
compiled variant, this behavior is resolved at compile time via operator
overloading.

This also affects smooth intersections.

I believe the behavior for the regular and smooth variants of union and inter
to be correct now.  Diff is correct for the case where RHS is unpainted, and I
would like to change the painted RHS case to allow material transfer, but the
current behavior is mostly fine.