WIP Occlusion Culling
The math isn't right yet.  This is most obvious on the "wedges" test model.
This is also missing the secondary stage for rendering the false negatives.

When enabled, this halves the frame time for most models, and I've found no
cases where this regresses performance.

This is a quick and dirty implementation that hacks the culling into the VS, so
models with lots of nodes will unfortunately still perform poorly.  But not as
bad as without!

There's still quite a bit of room to improve perf.  Switching to mesh shaders,
or implementing a compute stage to perform culling and set indirect draw
buffers would help substantially.  It would also help quite a bit if the
culling could be performed along the octree hierarchy, but that will need some
thought.  Octree/frustum culling could at least be performed on the CPU though.

Anyways, this is disabled by default for the time being.