Conditional Rendering & Occlusion Culling Debug Stuff
The purpose of this change is to introduce a conditional rendering mode
where the scene is only rendered and presented when on frames where SDL events
occurred, or when an effect requires it.  For example, the performance stats
overlay will force a redraw every frame until you disable it.  There is also a
"force redraw" option in the debug menu to accomplish the same thing.

This should greatly reduce waste heat while editing models, and is a nice
stopgap until more significant optimization work happens.  This will likely not
be useful for realtime games.

This change also introduces some harmless debug machinery for occlusion culling
that I wasn't planning on checking in, but I don't feel like untangling right
now, so I'll just remove it later when I'm done with it.